contributions user_id col

+++ api/internal/habit/contributions.go
@@ -9,6 +9,7 @@ import (
 
 type Contribution struct {
 	ID          int       `json:"id"`
+	UserID      int       `json:"userId" db:"user_id"`
 	Date        time.Time `json:"date"`
 	HabitID     int       `json:"habitId" db:"habit_id"`
 	Compeltions int       `json:"completions" db:"completions"`
@@ -19,25 +20,34 @@ type ContributionsRepo struct {
 	db *sqlx.DB
 }
 
+func NewContributionsRepo(db *sqlx.DB) *ContributionsRepo {
+	return &ContributionsRepo{
+		db: db,
+	}
+}
+
 type CreateContributionParams struct {
 	HabitID     int       `json:"habitId" db:"habit_id"`
+	UserID      int       `json:"userId" db:"user_id"`
 	Compeltions int       `json:"completions" db:"completions"`
 	Date        time.Time `json:"date" db:"date"`
 }
 
 type DeleteContributionParams struct {
 	HabitID int       `json:"habitId" db:"habit_id"`
+	UserID  int       `json:"userId" db:"user_id"`
 	Date    time.Time `json:"date" db:"date"`
 }
 
 type UpdateCompletionsParams struct {
 	ID          int `json:"id" db:"id"`
+	UserID      int `json:"userId" db:"user_id"`
 	Completions int `json:"completions" db:"completions"`
 }
 
-func (r *ContributionsRepo) List(habitID int) ([]Contribution, error) {
+func (r *ContributionsRepo) List(habitID int, userID int) ([]Contribution, error) {
 	contributions := []Contribution{}
-	err := r.db.Select(&contributions, "SELECT * FROM contributions WHERE habit_id=$1", habitID)
+	err := r.db.Select(&contributions, "SELECT * FROM contributions WHERE habit_id=$1 AND user_id = $2", habitID, userID)
 	if err != nil {
 		return nil, err
 	}
@@ -45,7 +55,7 @@ func (r *ContributionsRepo) List(habitID int) ([]Contribution, error) {
 }
 
 func (r *ContributionsRepo) Create(params CreateContributionParams) error {
-	query := "INSERT INTO contributions (habit_id, completions, date) VALUES (:habit_id, :completions, :date)"
+	query := "INSERT INTO contributions (habit_id, completions, date, user_id) VALUES (:habit_id, :completions, :date, :user_id)"
 	_, err := r.db.NamedExec(query, params)
 	if err != nil {
 		return err
@@ -57,9 +67,9 @@ func (r *ContributionsRepo) UpdateCompletions(params UpdateCompletionsParams) er
 	query := `
 		UPDATE contributions
 	  SET completions = $1 
-		WHERE id = $2;
+		WHERE id = $2 AND user_id = $3;
 	`
-	_, err := r.db.Exec(query, params.Completions, params.ID)
+	_, err := r.db.Exec(query, params.Completions, params.ID, params.UserID)
 	if err != nil {
 		return err
 	}
@@ -67,16 +77,10 @@ func (r *ContributionsRepo) UpdateCompletions(params UpdateCompletionsParams) er
 	return nil
 }
 
-func (r *ContributionsRepo) Delete(id int) error {
-	_, err := r.db.Exec("DELETE FROM contributions WHERE id=$1", id)
+func (r *ContributionsRepo) Delete(id int, userID int) error {
+	_, err := r.db.Exec("DELETE FROM contributions WHERE id=$1 AND user_id = $2", id, userID)
 	if err != nil {
 		return err
 	}
 	return nil
-}
-
-func NewContributionsRepo(db *sqlx.DB) *ContributionsRepo {
-	return &ContributionsRepo{
-		db: db,
-	}
 }
